% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bodyfat}
\alias{bodyfat}
\title{Body fat data set}
\format{A data frame with 128 observations on 15 variables:
\describe{
\item{list("age")}{Age}
\item{list("sex")}{Gender}
\item{list("bmi")}{Body mass index}
\item{list("map")}{Mean arterial pressure (average blood pressure)}
\item{list("tc")}{Total cholesterol (mg/dL)? Desirable range: below 200 mg/dL}
\item{list("ldl")}{Low-density lipoprotein ("bad" cholesterol)? Desirable range: below 130 mg/dL }
\item{list("hdl")}{High-density lipoprotein ("good" cholesterol)? Desirable range: above 40 mg/dL}
\item{list("tch")}{Blood serum measurement}
\item{list("ltg")}{Blood serum measurement}
\item{list("glu")}{Blood serum measurement (glucose?)}
\item{list("y")}{A quantitative measure of disease progression one year after baseline}
}}
\usage{
data(bodyfat)
}
\description{
A data frame with 128 observations on 15 variables.
}
\details{
Details about the bodyfat data set go here.
}
\examples{
data(bodyfat)
full.mod = lm(Bodyfat~.,data=subset(bodyfat,select=-Id))
}
\references{
Johnson W (1996, Vol 4). Fitting percentage of
  body fat to simple body measurements. Journal of Statistics
  Education. Bodyfat data retrieved from
  http://www.amstat.org/publications/jse/v4n1/datasets.johnson.html
  An expanded version is included in the \code{mfp} R package.
}
\keyword{datasets}

