% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fence}
\alias{fence}
\title{Fence based model selection}
\usage{
fence(fixed, random, data, method = c("ML", "MVC"), adaptive = TRUE,
  B = 100, R = 100, nvmax, trace = FALSE, exhaustive = FALSE,
  use.leaps = FALSE, normalize = TRUE, intercept = TRUE)
}
\arguments{
\item{fixed}{formula specifying the full fixed model}

\item{random}{(optional) formula specifying the full random model
using the syntax of lme() #MAY CHANGE TO lme4()#}

\item{data}{data frame containing all (fixed and random) variables}

\item{method}{the model selection method to be used. Currently
only \code{method = "ML"} is supported (perhaps in the future
\code{method = "MVC"} will be implemented).}

\item{adaptive}{logical. If \code{TRUE} the boundary of the fence is
given by cstar.  Otherwise, it the original (non-adaptive) fence
is performed where the boundary is cstar*hat(sigma)_{M,tildeM}.}

\item{B}{number of bootstrap replications for the adaptive fence}
}
\description{
A wrapper function that combines \code{lmfence} (or, in the
future \code{lmmfence}) and \code{af} that performs the adaptive
fence procedure in an automated way, then extracts the optimal model
and presents diagnostics.
Implements the fence of Jiang (2008)
Assumes number of variables << number of observations
and that the full model is estimable
also currently assumes the random effects are known
i.e. only does model selection for fixed effects
}

