% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{fencer-package}
\alias{fencer-package}
\alias{granovaGG}
\title{Model selection based on the fence procedure.}
\description{
Model selection based on the fence procedure
}
\details{
More detail
...
}
\author{
Garth Tarr
}
\seealso{
\code{\link{...}}
}
\keyword{package}

