% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lmmfence}
\alias{lmmfence}
\title{The fence procedure for linear mixed models}
\usage{
lmmfence(mf, cstar, nvmax, method = "ML", adaptive = TRUE, trace = TRUE,
  force.in = NULL, ...)
}
\description{
Relies on a varient of the wild bootstrap in that
it only considers the fixed effects.
}

