% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lvp}
\alias{lvp}
\title{Model selection/stability curves and variable importance}
\usage{
lvp(mf, nvmax, B = 100, lambda.max, ...)
}
\arguments{
\item{mf}{the full fitted model of type lm or glm}

\item{nvmax}{the max number of variables to consider as being important}

\item{B}{the number of bootstrap replications}

\item{lambda.max}{the maximum penalty value for the vip plot, defaults to log(n)}
}
\description{
Performs the heavy lifting and provides the plot methods for msc, boot and vip
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

