% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mextract}
\alias{mextract}
\title{Extract model elements}
\usage{
mextract(model, screen)
}
\arguments{
\item{model}{a fitted 'full' model, the result of a call
to lm or glm (and in the future lme or lmer).}
}
\description{
This function extracts things like the formula,
data matrix, etc. from a lm or glm object
}

