% Generated by roxygen2 (4.0.1): do not edit by hand
\name{msc}
\alias{msc}
\title{Model selection/stability curves}
\usage{
msc(fixed, data, family, lvp, fence = FALSE, B = 100, ...)
}
\arguments{
\item{data}{a data frame containing the variables in the full
model.  Currently this is required.  Future development will
look at the \code{lm} function to replicate functionality,
whereby if \code{data} is missing or variables are not found
in \code{data}, the variables are taken from
\code{environment(formula)}, typically the
environment from which \code{lmfence} is called.}
}
\description{
Opens a shiny GUI to investigate model selection/stability issues
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

