% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.af}
\alias{plot.af}
\title{Plot diagnostics for an af object}
\usage{
\method{plot}{af}(x, pch, ...)
}
\arguments{
\item{x}{\code{af} object, the result of \code{\link{af}}}

\item{...}{other parameters to be passed through to
plotting functions.}
}
\description{
Summary plot of the bootstrap results of an
af object.
}

