% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bglmnet.R
\name{plot.bglmnet}
\alias{plot.bglmnet}
\title{Plot diagnostics for a bglmnet object}
\usage{
\method{plot}{bglmnet}(x, highlight, classic = FALSE, html.only = FALSE,
  which = c("models", "variables"), width = 800, height = 400,
  fontSize = 12, left = 50, top = 30, chartWidth = "60\%",
  chartHeight = "80\%", axisTitlesPosition = "out", dataOpacity = 0.5,
  options = NULL, shiny = FALSE, backgroundColor = "transparent",
  plb = 0.01, ...)
}
\arguments{
\item{x}{\code{bglmnet} object, the result of \code{\link{bglmnet}}}

\item{highlight}{the name of a variable that will be highlighted.}

\item{classic}{logical.  If \code{classic=TRUE} a
base graphics plot is provided instead of a googleVis plot.
Default is \code{classic=FALSE}.}

\item{html.only}{logical. Use \code{html.only=TRUE} when including
interactive plots in markdown documents (this includes rpres files).}

\item{which}{a vector specifying the plots to be output. Variable
inclusion type plots \code{which="variables"} or model description loss against
penalty parameter \code{which="models"}.}

\item{width}{Width of the googleVis chart canvas area, in pixels.
Default: 800.}

\item{height}{Height of the googleVis chart canvas area, in pixels.
Default: 400.}

\item{fontSize}{font size used in googleVis chart.  Default: 12.}

\item{left}{space at left of chart (pixels?).  Default: "50".}

\item{top}{space at top of chart (pixels?).  Default: "30".}

\item{chartWidth}{googleVis chart area width.
A simple number is a value in pixels;
a string containing a number followed by \code{\%} is a percentage.
Default: \code{"60\%"}}

\item{chartHeight}{googleVis chart area height.
A simple number is a value in pixels;
a string containing a number followed by \code{\%} is a percentage.
Default: \code{"80\%"}}

\item{axisTitlesPosition}{Where to place the googleVis axis titles,
compared to the chart area. Supported values:
"in" - Draw the axis titles inside the the chart area.
"out" - Draw the axis titles outside the chart area.
"none" - Omit the axis titles.}

\item{dataOpacity}{The transparency of googleVis data points,
with 1.0 being completely opaque and 0.0 fully transparent.}

\item{options}{a list to be passed to the googleVis function giving
complete control over the output.  Specifying a value for
\code{options} overwrites all other plotting variables.}

\item{shiny}{logical. Used internally to facilitate proper display
of plots within the mplot shiny user interface.  Use
\code{shiny=TRUE} when displaying output within a shiny interface.}

\item{backgroundColor}{The background colour for the main area
of the chart. A simple HTML color string,
for example: 'red' or '#00cc00'.  Default: 'transparent'}

\item{plb}{lower bound on the probability of a model being selected. If
a model has a selection probability lower than plb it will not be
plotted.}

\item{...}{further arguments (currently unused)}
}
\description{
A plot method to visualise the results of a \code{bglmnet} object.
}
\examples{
n = 100
set.seed(11)
e = rnorm(n)
x1 = rnorm(n)
x2 = rnorm(n)
x3 = x1^2
x4 = x2^2
x5 = x1*x2
y = 1 + x1 + x2 + e
dat = data.frame(y,x1,x2,x3,x4,x5)
lm1 = lm(y~.,data=dat)
bgn1 = mplot:::bglmnet(lm1)
\dontrun{
plot(bgn1,highlight="x1")
}
}
\seealso{
\code{\link{bglmnet}}
}

