% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.lvp}
\alias{plot.lvp}
\title{Plot diagnostics for an lvp object}
\usage{
\method{plot}{lvp}(x, highlight, classic = FALSE, html.only = FALSE, ...)
}
\arguments{
\item{x}{\code{lvp} object, the result of \code{\link{lvp}}}

\item{highlight}{a vector indicating which variables should be highlighted.}

\item{classic}{logical.  If \code{classic=TRUE} a
base graphics plot is provided instead of a googleVis plot.
Default is \code{classic=FALSE}.}

\item{...}{other parameters to be passed through to
plotting functions.}
}
\description{
Summary plot of the lvp results.
}

