% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.vis}
\alias{plot.vis}
\title{Plot diagnostics for an vis object}
\usage{
\method{plot}{vis}(x, highlight, classic = FALSE, html.only = FALSE,
  which = c("vip", "lvk", "boot"), width = 800, height = 400,
  fontSize = 12, left = 50, top = 30, chartWidth = "60\%",
  chartHeight = "80\%", axisTitlesPosition = "out", dataOpacity = 0.5,
  options = NULL, ...)
}
\arguments{
\item{x}{\code{vis} object, the result of \code{\link{vis}}}

\item{highlight}{a vector indicating which variables
should be highlighted.}

\item{classic}{logical.  If \code{classic=TRUE} a
base graphics plot is provided instead of a googleVis plot.
Default is \code{classic=FALSE}.}

\item{...}{other parameters to be passed through to
plotting functions.}
}
\description{
Summary plot of the vis results.
}

