% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.R
\name{print.af}
\alias{print.af}
\title{Print method for an af object}
\usage{
\method{print}{af}(x, best.only = TRUE, ...)
}
\arguments{
\item{x}{an \code{af} object, the result of \code{\link{af}}}

\item{best.only}{logical determining whether the output used the
standard fence approach of only considering the best models
that pass the fence (\code{TRUE}) or if it should take into
account all models that pass the fence at each boundary
value (\code{FALSE}).}

\item{...}{further arguments (currently unused)}
}
\description{
Prints basic output of the bootstrap results of an
af object.
}
