% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{print.vis}
\alias{print.vis}
\title{Print method for a vis object}
\usage{
\method{print}{vis}(x, min.prob = 0.3, print.full.model = FALSE, ...)
}
\arguments{
\item{x}{a \code{vis} object, the result of \code{\link{vis}}}

\item{min.prob}{a lower bound on the probability of
selection before the result is printed}

\item{print.full.model}{logical, determines if the full
model gets printed or not.  Default=FALSE.}

\item{...}{further arguments (currently unused)}
}
\description{
Prints basic output of the bootstrap results of an
vis object.
}

