% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot-package.R
\name{process.fn}
\alias{process.fn}
\title{Process results within af function}
\usage{
process.fn(fence.mod, fence.rank)
}
\arguments{
\item{fence.mod}{set of fence models}

\item{fence.rank}{set of fence model ranks}
}
\description{
This function is used by the af function to process
the results when iterating over different boundary values
}
\keyword{internal}
