% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Qm.R
\name{Qm}
\alias{Qm}
\title{A measure of lack of fit}
\usage{
Qm(object, method)
}
\arguments{
\item{object}{Any object from which a log-likelihood value can be extracted.}

\item{method}{method the model selection method to be used. Currently
only \code{method = "ML"} is supported (perhaps in the future
\code{method = "MVC"} will be implemented).}
}
\description{
This function calculates a lack-of-fit measure for
a model, \eqn{\hat{Q}_M}.  It currently simply
returns the negative log-likelihood value of an
estimated model object.
}

