% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplot-package.R
\name{safeDeparse}
\alias{safeDeparse}
\title{Safe deparse}
\usage{
safeDeparse(expr)
}
\arguments{
\item{expr}{expression to be safely deparsed}
}
\description{
Supports long formula construction
}

