% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sigMM}
\alias{sigMM}
\title{Standard deviation of hatQ_M and hatQ_tildeM}
\usage{
sigMM(k.mod, method, k.full, adaptive)
}
\arguments{
\item{k.mod}{number of parameters in the estimated model}

\item{method}{method the model selection method to be used. Currently
only \code{method = "ML"} is supported (perhaps in the future
\code{method = "MVC"} will be implemented).}

\item{k.full}{number of parameters in the full model}

\item{adaptive}{logical. If \code{TRUE} the boundary of the fence is
given by cstar.  Otherwise, it the original (non-adaptive) fence
is performed where the boundary is cstar*hat(sigma)_{M,tildeM}.}
}
\description{
This function calculates \eqn{\hat(sigma)_{M,tildeM}}
 the standard deviation of
the difference between the two lack-of-fit measures
\eqn{\hat{Q}_M} and \eqn{\hat{Q}_{\tilde{M}}} as
described in Jiang et. al. (2008).  When using the
adaptive fence procedure, this quantity no longer needs
to be calculated and simply returns a value of 1.
}

