% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bglmnet.R
\name{bglmnet}
\alias{bglmnet}
\title{Boostrap model selection plots for glmnet}
\usage{
bglmnet(mf, nlambda = NULL, lambda = seq(0.05, 0.95, 0.05), B = 100,
  probaseuil = 1, penalty.factor, random, screen = FALSE)
}
\arguments{
\item{mf}{a fitted 'full' model, the result of a call
to lm or glm (and in the future lme or lmer).}

\item{nlambda}{how many penalty values to consider.}

\item{B}{number of bootstrap replications}

\item{n.cores}{number of cores to be used when parallel
processing the bootstrap (Not yet implemented.)}

\item{force.in}{the names of variables that should be forced
into all estimated models. (Not yet implemented.)}

\item{...}{further arguments (currently unused)}
}
\description{
Experimental!!
}
\details{
The result of this function is essentially just a
  list. The supplied plot method provides a way to visualise the
  results.
}
\examples{
n = 100
set.seed(11)
e = rnorm(n)
x1 = rnorm(n)
x2 = rnorm(n)
x3 = x1^2
x4 = x2^2
x5 = x1*x2
y = 1 + x1 + x2 + e
dat = data.frame(y,x1,x2,x3,x4,x5)
lm1 = lm(y~.,data=dat)
bgn1 = mplot:::bglmnet(lm1)
\dontrun{
plot(bgn1,highlight="x1")
}
}
\seealso{
\code{\link{plot.bglmnet}}
}

