% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bodyfat}
\alias{bodyfat}
\title{Body fat data set}
\format{A data frame with 128 observations on 15 variables.}
\usage{
data(bodyfat)
}
\description{
A data frame with 128 observations on 15 variables.
}
\details{
Details about the bodyfat data set go here.
}
\examples{
data(bodyfat)
full.mod = lm(Bodyfat~.,data=subset(bodyfat,select=-Id))
af.bf = af(full.mod,n.cores=detectCores(),initial.stepwise=TRUE)
plot(af.bf)
summary(af.bf)
af.bf2 = af(full.mod,n.cores=detectCores(),c.max=20)
plot(af.bf2)
mplot(full.mod,af = af.bf2)
}
\references{
Johnson W (1996, Vol 4). Fitting percentage of
  body fat to simple body measurements. Journal of Statistics
  Education. Bodyfat data retrieved from
  http://www.amstat.org/publications/jse/v4n1/datasets.johnson.html
  An expanded version is included in the \code{mfp} R package.
}
\keyword{datasets}

