% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bodyfat}
\alias{bodyfat}
\title{Body fat data set}
\format{A data frame with 128 observations on 15 variables.}
\usage{
data(bodyfat)
}
\description{
Description of the body fat data set here.  This is more
an example to show how I can document a real data set included
with the package.  I don't think this is the best real data set
to use.  See the documentation in \code{\link{af}} for an
example which uses it.
}
\details{
Details about the bodyfat data set go here.
}
\references{
Johnson W (1996, Vol 4). Fitting percentage of
  body fat to simple body measurements. Journal of Statistics
  Education. Bodyfat data retrieved from
  http://www.amstat.org/publications/jse/v4n1/datasets.johnson.html
  An expanded version is included in the \code{mfp} R package.
}
\keyword{datasets}

