% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Blood and other measurements in diabetics}
\format{A data frame with 442 observations on 11 variables.}
\usage{
data(diabetes)
}
\description{
The diabetes data frame has 442 rows and 11 columns.
These are the data used in Efron et al. (2003).
}
\details{
Data sourced from http://web.stanford.edu/~hastie/Papers/LARS
}
\examples{
data(diabetes)
full.mod = lm(y~.,data=diabetes)
}
\references{
Efron, B., Hastie, T., Johnstone, I., Tibshirani, R., (2004).
  Least angle regression. The Annals of Statistics 32(2) 407-499.
  DOI: 10.1214/009053604000000067
}
\keyword{datasets}

