% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmfence}
\alias{glmfence}
\title{The fence procedure for generalised linear models}
\usage{
glmfence(mf, cstar, nvmax, method = "ML", adaptive = TRUE, trace = TRUE,
  ...)
}
\arguments{
\item{mf}{an object of class \code{\link[stats]{glm}}
specifying the full model.}

\item{cstar}{the boundary of the fence, typically found
through bootstrapping.}

\item{nvmax}{the maximum number of variables that will be
be considered in the model (the default is the number of
explanatory variables in \code{fixed}).}

\item{method}{the model selection method to be used. Currently
only \code{method = "ML"} is supported (perhaps in the future
\code{method = "MVC"} will be implemented).}

\item{adaptive}{logical. If \code{TRUE} the boundary of the fence is
given by cstar.  Otherwise, it the original (non-adaptive) fence
is performed where the boundary is cstar*hat(sigma)_{M,tildeM}.}

\item{trace}{logical. If \code{TRUE} the function prints out its
progress as it iterates up through the dimensions.}

\item{...}{optional extra parameters passed through (?? necessary)}
}
\description{
This function implements the fence procedure to
find the best generalised linear model.
}
\references{
Tarr, Müller and Welsh (2014)
}
\seealso{
\code{\link{adaptfence}}, \code{\link{lmmfence}}

Other fence: \code{\link{af}}; \code{\link{lmfence}}
}

