% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lvp}
\alias{lvp}
\title{Model selection/stability curves and variable importance}
\usage{
lvp(fixed, data, family, nvmax, B = 100, lambda.max, ...)
}
\arguments{
\item{fixed}{}

\item{data}{a data frame containing the variables in the full
model.  Currently this is required.  Future development will
look at the \code{lm} function to replicate functionality,
whereby if \code{data} is missing or variables are not found
in \code{data}, the variables are taken from
\code{environment(formula)}, typically the
environment from which the function is called.}

\item{family}{for glms}

\item{nvmax}{the max number of variables to consider as being important}

\item{B}{the number of bootstrap replications}

\item{lambda.max}{the maximum penalty value for the vip plot, defaults to log(n)}
}
\description{
Performs the heavy lifting and provides the plot methods for msc, boot and vip
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

