% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplot-package.R
\name{mextract}
\alias{mextract}
\title{Extract model elements}
\usage{
mextract(model, screen = FALSE, redundant = TRUE)
}
\arguments{
\item{model}{a fitted 'full' model, the result of a call
to lm or glm (and in the future lme or lmer).}

\item{screen}{logical, whether or not to perform an initial
screen for outliers.  Highly experimental, use at own risk.
Default = FALSE.}

\item{redundant}{logical, whether or not to add a redundant
variable.  Default = TRUE.}
}
\description{
This function extracts things like the formula,
data matrix, etc. from a lm or glm object
}

