% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot-package.R
\docType{package}
\name{mplot-package}
\alias{mplot-package}
\title{Graphical model stability and model selection procedures}
\description{
Graphical model stability and model selection procedures
}
\keyword{package}
