% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot-package.R
\docType{package}
\name{mplot-package}
\alias{mplot-package}
\title{Graphical model stability and model selection procedures}
\description{
Graphical model stability and model selection procedures
}
\references{
Tarr G, Mueller S and Welsh AH (2018). mplot: An R Package for
Graphical Model Stability and Variable Selection Procedures.
Journal of Statistical Software, 83(9), pp. 1-28. doi: 10.18637/jss.v083.i09
}
\keyword{package}
