% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mplot-package}
\alias{mplot-package}
\title{Graphical model stability and model selection procedures}
\description{
Graphical model stability and model selection procedures
}
\details{
...
}
\author{
Garth Tarr
}
\seealso{
\code{\link{...}}
}
\keyword{package}

