% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mplot}
\alias{mplot}
\title{Model selection and stability curves}
\usage{
mplot(mf, vis = NULL, af = NULL, B = 100, ...)
}
\arguments{
\item{mf}{a fitted model.}

\item{vis}{(optional) an object of type vis created using the
vis() function.}

\item{af}{(optional) an af object of}
}
\description{
Opens a shiny GUI to investigate a range of model selection
and stability issues
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

