% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.af}
\alias{print.af}
\title{Print method for an af object}
\usage{
\method{print}{af}(x, ...)
}
\arguments{
\item{x}{\code{af} object, the result of \code{\link{af}}}

\item{...}{other parameters to be passed through to
plotting functions.}
}
\description{
Prints basic output of the bootstrap results of an
af object.
}

