% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process.fn}
\alias{process.fn}
\title{Process results within af function}
\usage{
process.fn(fence.mod, fence.rank)
}
\description{
This function is used by the af function to process
the results when iterating over different boundary values
}

