% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Qm}
\alias{Qm}
\title{A measure of lack of fit}
\usage{
Qm(object, method)
}
\arguments{
\item{object}{an object of class \code{"lm"}.  Actually any
object from which a log-likelihood value can be extracted.
Will need more thought if/when other methods are implemented.}

\item{method}{method the model selection method to be used. Currently
only \code{method = "ML"} is supported (perhaps in the future
\code{method = "MVC"} will be implemented).}
}
\description{
This function calculates a lack-of-fit measure for
a model, \eqn{\hat{Q}_M}.  It currently simply
returns the negative log-likelihood value of an
estimated model object.
}

