% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qrange.R
\name{qrange}
\alias{qrange}
\title{Identify an appropriate rance of values over which to bootstrap}
\usage{
qrange(k.range, yname, fixed, data, method, force.in, model.type, family)
}
\arguments{
\item{k.range}{list with dimension elements k.max and k.min}

\item{yname}{name of dependent variable}

\item{fixed}{the full model formula}

\item{data}{full data table}

\item{method}{method used in Qm}

\item{force.in}{which variables to force into the model}

\item{model.type}{currently only lm or glm}

\item{family}{for glms.}
}
\description{
This function takes an upper and lower dimension size (obtained by
forwards and backwards model selection and then adding and subtracting
2 from each of the extremes to encompas a broader range of models).
For both the small and large model size, the "best" model is identified
using the \code{leaps} package and the corresponding lack of fit measure
is calculated.
}

