% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qrange}
\alias{qrange}
\title{Identify an appropriate rance of values over which to bootstrap}
\usage{
qrange(k.range, yname, fixed, data, method, force.in)
}
\description{
This function takes an upper and lower dimension size (obtained by
forwards and backwards model selection and then adding and subtracting
1 from each of the extremes to encompas a broader range of models).
For both the small and large model size, the "best" model is identified
using the \code{leaps} package and the corresponding lack of fit measure
is calculated.
}

