% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/af.R
\name{summary.af}
\alias{summary.af}
\title{Summary method for an af object}
\usage{
summary.af(object, best.only = TRUE, ...)
}
\arguments{
\item{object}{\code{af} object, the result of \code{\link{af}}}

\item{best.only}{logical determining whether the output used the
standard fence approach of only considering the best models
that pass the fence (\code{TRUE}) or if it should take into
account all models that pass the fence at each boundary
value (\code{FALSE}).}

\item{...}{further arguments (currently unused)}
}
\description{
Provides comprehensive  output of the bootstrap results of an
af object.
}

