% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot-package.R
\name{txt.fn}
\alias{txt.fn}
\title{Print text for fence methods}
\usage{
txt.fn(score, UB, obj)
}
\arguments{
\item{score}{realised value}

\item{UB}{upper bound}

\item{obj}{fitted model object}
}
\description{
This function provides the text for the case when trace=TRUE
when using lmfence and glmfence functions.
}
\keyword{internal}

