% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vis}
\alias{vis}
\title{Model selection and stability curves}
\usage{
vis(mf, lvp = NULL, af = NULL, B = 100, ...)
}
\arguments{
\item{mf}{a fitted model.}

\item{lvp}{(optional) an object of type lvp created using the
lvp() function.}

\item{af}{(optional) an af object of}
}
\description{
Opens a shiny GUI to investigate a range of model selection
and stability issues
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

