% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vis}
\alias{vis}
\title{Model selection/stability curves and variable importance}
\usage{
vis(mf, nvmax, B = 100, lambda.max, ...)
}
\arguments{
\item{mf}{the full fitted model of type lm or glm}

\item{nvmax}{the max number of variables to consider as being important}

\item{B}{the number of bootstrap replications}

\item{lambda.max}{the maximum penalty value for the vip plot,
defaults to 2*log(n)}
}
\description{
Performs the heavy lifting and provides the plot methods for lvk,
boot and vip
}
\references{
Müller and Welsh (2010); Murray, Heritier and Müller (2013)
}

