% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{Model stability curves and variable inclusion plots}
\usage{
vis(mf, nvmax, B = 100, lambda.max, n.cores = 2, force.in = NULL,
  screen = FALSE, ...)
}
\arguments{
\item{mf}{a fitted 'full' model, the result of a call
to lm or glm (and in the future lme or lmer).}

\item{nvmax}{size of the largest model that can still be
considered as a viable candidate.}

\item{B}{number of bootstrap replications}

\item{lambda.max}{maximum penalty value for the vip plot,
defaults to 2*log(n)}

\item{n.cores}{number of cores to be used when parallel
processing the bootstrap.}

\item{force.in}{the names of variables that should be forced
into all estimated models. (Not yet implemented.)}

\item{...}{further arguments (currently unused)}
}
\description{
Calculates and provides the plot methods for standard
and bootstrap enhanced model stability curves (lvk and
boot) as well as variable inclusion plots (vip).
}
\details{
The result of this function is essentially just a
  list. The supplied plot method provides a way to visualise the
  results.
}
\examples{
n = 100
set.seed(11)
e = rnorm(n)
x1 = rnorm(n)
x2 = rnorm(n)
x3 = x1^2
x4 = x2^2
x5 = x1*x2
y = 1 + x1 + x2 + e
dat = data.frame(y,x1,x2,x3,x4,x5)
lm1 = lm(y~.,data=dat)
v1 = vis(lm1)
\dontrun{
plot(v1,highlight="x1")
}
}
\references{
Mueller, S. and Welsh, A. H. (2010), On Model
  Selection Curves. International Statistical Review, 78:240-256.
  doi: 10.1111/j.1751-5823.2010.00108.x

  Murray, K., Heritier, S. and Mueller, S. (2013), Graphical
  tools for model selection in generalized linear models.
  Statistics in Medicine, 32:4438-4451. doi: 10.1002/sim.5855
}
\seealso{
\code{\link{plot.vis}}
}

